+++ launch/mcmh_localization.launch
@@ -19,6 +19,8 @@
   <!-- Nó do estimador baseado em MCMC-->
   <node name="mcmh_localizer" pkg="mcmh_localization" type="mcmh_localizer.py" output="screen" /> 
 
+  <node name="pose_broadcaster" pkg="mcmh_localization" type="pose_broadcaster.py" output="screen" /> 
+
   <!-- RViz com visualização do mapa, robot e partículas -->
   <node pkg="rviz" type="rviz" name="rviz" args="-d $(find mcmh_localization)/rviz/mcmh_view.rviz" />
 </launch>
